+++ pwa/firebase-config.js
@@ -1,0 +1,8 @@
+window.I_EXPENSE_FIREBASE_CONFIG = window.I_EXPENSE_FIREBASE_CONFIG || {
+  apiKey: "",
+  authDomain: "",
+  projectId: "",
+  appId: "",
+  messagingSenderId: "",
+  storageBucket: ""
+};